refactor: update buildQuote implementation and remove deprecated adapters
