+++ .changeset/warm-owls-build.md
@@ -9,4 +9,3 @@ feat: add buildQuote for on-chain intent submission without a solver API
 -   Implement `buildQuote` in `AcrossProvider` using `SpokePool.deposit()` calldata encoding
 -   Add `BuildQuoteRequest` schema with required amounts, escrow contract address, and fill deadline
 -   Expose `buildQuote` on the `Aggregator` for provider-routed local quote building
--   Export `OPEN_ABI`, `BuildQuoteRequest`, `buildOifQuote`, and `buildAcrossQuote` from the public API
