+++ packages/cross-chain/src/protocols/across/index.ts
@@ -6,4 +6,3 @@ export * from "./schemas.js";
 export * from "./discoverySchemas.js";
 export * from "./constants.js";
 export * from "./types.js";
-export * from "./buildQuoteAdapter.js";
