+++ packages/cross-chain/test/adapters/buildAcrossQuoteAdapter.spec.ts
@@ -4,11 +4,11 @@ import { baseSepolia, sepolia } from "viem/chains";
 import { describe, expect, it } from "vitest";
 
 import type { BuildQuoteRequest } from "../../src/core/schemas/quoteRequest.js";
-import { buildAcrossQuote } from "../../src/protocols/across/buildQuoteAdapter.js";
 import {
     ACROSS_SPOKE_POOL_ADDRESSES,
     ACROSS_SPOKE_POOL_DEPOSIT_ABI,
 } from "../../src/protocols/across/constants.js";
+import { AcrossProvider } from "../../src/protocols/across/provider.js";
 
 const USER = "0x742D35cC6634C0532925A3b844bc9E7595f0BEb8" as Address;
 const RECIPIENT = "0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045" as Address;
@@ -16,6 +16,8 @@ const INPUT_TOKEN = "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48" as Address;
 const OUTPUT_TOKEN = "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913" as Address;
 const FALLBACK_SPOKE_POOL = "0x95AD61B0A150D79219dcf64E1e6cC01f0C0c8A4A" as Address;
 const PROVIDER_ID = "test-across";
+
+const provider = new AcrossProvider({ providerId: PROVIDER_ID });
 
 function createRequest(overrides?: Partial<BuildQuoteRequest>): BuildQuoteRequest {
     return {
@@ -36,16 +38,16 @@ function createRequest(overrides?: Partial<BuildQuoteRequest>): BuildQuoteReques
     };
 }
 
-describe("buildAcrossQuote", () => {
-    it("returns a Quote with a single TransactionStep", () => {
-        const quote = buildAcrossQuote(createRequest(), PROVIDER_ID);
+describe("AcrossProvider.buildQuote", () => {
+    it("returns a Quote with a single TransactionStep", async () => {
+        const quote = await provider.buildQuote(createRequest());
 
         expect(quote.order.steps).toHaveLength(1);
         expect(quote.order.steps[0]!.kind).toBe("transaction");
     });
 
-    it("uses known SpokePool address for supported chains", () => {
-        const quote = buildAcrossQuote(createRequest(), PROVIDER_ID);
+    it("uses known SpokePool address for supported chains", async () => {
+        const quote = await provider.buildQuote(createRequest());
         const step = quote.order.steps[0]!;
 
         expect(step.kind).toBe("transaction");
@@ -54,12 +56,11 @@ describe("buildAcrossQuote", () => {
         }
     });
 
-    it("falls back to escrowContractAddress for unknown chains", () => {
-        const quote = buildAcrossQuote(
+    it("falls back to escrowContractAddress for unknown chains", async () => {
+        const quote = await provider.buildQuote(
             createRequest({
                 input: { chainId: 999999, assetAddress: INPUT_TOKEN, amount: "1000000" },
             }),
-            PROVIDER_ID,
         );
         const step = quote.order.steps[0]!;
 
@@ -69,7 +70,7 @@ describe("buildAcrossQuote", () => {
         }
     });
 
-    it("encodes valid deposit() calldata with correct args", () => {
+    it("encodes valid deposit() calldata with correct args", async () => {
         const params = createRequest({
             output: {
                 chainId: baseSepolia.id,
@@ -78,7 +79,7 @@ describe("buildAcrossQuote", () => {
                 recipient: RECIPIENT,
             },
         });
-        const quote = buildAcrossQuote(params, PROVIDER_ID);
+        const quote = await provider.buildQuote(params);
         const step = quote.order.steps[0]!;
 
         expect(step.kind).toBe("transaction");
@@ -126,9 +127,9 @@ describe("buildAcrossQuote", () => {
         }
     });
 
-    it("sets quoteTimestamp to a recent value", () => {
+    it("sets quoteTimestamp to a recent value", async () => {
         const before = Math.floor(Date.now() / 1000);
-        const quote = buildAcrossQuote(createRequest(), PROVIDER_ID);
+        const quote = await provider.buildQuote(createRequest());
         const after = Math.floor(Date.now() / 1000);
 
         const step = quote.order.steps[0]!;
@@ -144,8 +145,8 @@ describe("buildAcrossQuote", () => {
         }
     });
 
-    it("uses user as recipient when no explicit recipient", () => {
-        const quote = buildAcrossQuote(createRequest(), PROVIDER_ID);
+    it("uses user as recipient when no explicit recipient", async () => {
+        const quote = await provider.buildQuote(createRequest());
 
         expect(quote.preview.outputs[0]!.accountAddress).toBe(USER);
 
@@ -162,9 +163,9 @@ describe("buildAcrossQuote", () => {
         }
     });
 
-    it("populates preview.inputs correctly", () => {
+    it("populates preview.inputs correctly", async () => {
         const params = createRequest();
-        const quote = buildAcrossQuote(params, PROVIDER_ID);
+        const quote = await provider.buildQuote(params);
 
         expect(quote.preview.inputs).toHaveLength(1);
         expect(quote.preview.inputs[0]!.chainId).toBe(sepolia.id);
@@ -172,9 +173,9 @@ describe("buildAcrossQuote", () => {
         expect(quote.preview.inputs[0]!.amount).toBe("1000000");
     });
 
-    it("populates preview.outputs correctly", () => {
+    it("populates preview.outputs correctly", async () => {
         const params = createRequest();
-        const quote = buildAcrossQuote(params, PROVIDER_ID);
+        const quote = await provider.buildQuote(params);
 
         expect(quote.preview.outputs).toHaveLength(1);
         expect(quote.preview.outputs[0]!.chainId).toBe(baseSepolia.id);
@@ -182,9 +183,9 @@ describe("buildAcrossQuote", () => {
         expect(quote.preview.outputs[0]!.amount).toBe("980000");
     });
 
-    it("populates order.checks.allowances", () => {
+    it("populates order.checks.allowances", async () => {
         const params = createRequest();
-        const quote = buildAcrossQuote(params, PROVIDER_ID);
+        const quote = await provider.buildQuote(params);
 
         expect(quote.order.checks!.allowances).toHaveLength(1);
         const allowance = quote.order.checks!.allowances![0]!;
@@ -195,22 +196,22 @@ describe("buildAcrossQuote", () => {
         expect(allowance.required).toBe("1000000");
     });
 
-    it("sets the provider field", () => {
-        const quote = buildAcrossQuote(createRequest(), PROVIDER_ID);
+    it("sets the provider field", async () => {
+        const quote = await provider.buildQuote(createRequest());
         expect(quote.provider).toBe(PROVIDER_ID);
     });
 
-    it("includes buildQuote metadata", () => {
+    it("includes buildQuote metadata", async () => {
         const params = createRequest();
-        const quote = buildAcrossQuote(params, PROVIDER_ID);
+        const quote = await provider.buildQuote(params);
 
         expect(quote.metadata!.buildQuote).toBe(true);
         expect(quote.metadata!.fillDeadline).toBe(params.fillDeadline);
         expect(quote.metadata!.spokePoolAddress).toBe(ACROSS_SPOKE_POOL_ADDRESSES[sepolia.id]);
     });
 
-    it("sets the correct origin chainId on the step", () => {
-        const quote = buildAcrossQuote(createRequest(), PROVIDER_ID);
+    it("sets the correct origin chainId on the step", async () => {
+        const quote = await provider.buildQuote(createRequest());
         expect(quote.order.steps[0]!.chainId).toBe(sepolia.id);
     });
 });
